Rename bolt_draw BoltDrawManager and extract to separate file.
