+++ engine/src/star_system.h
@@ -118,7 +118,7 @@ public:
     Statistics stats;
     std::multimap<Unit*, Unit*> last_collisions;
     CollideMap *collide_map[2]; // 0 Unit 1 Bolt
-    class bolt_draw *bolts = nullptr;
+    class BoltDrawManager *bolts = nullptr;
     class CollideTable *collide_table = nullptr;
 
 protected:
